+++ examples/bevy_state.rs
@@ -3,7 +3,7 @@
 use bevy::{input::common_conditions::input_just_pressed, prelude::*};
 use pyri_state::prelude::*;
 
-use iyes_progress::prelude::*;
+//use iyes_progress::prelude::*;
 
 fn main() -> AppExit {
     App::new()
@@ -13,11 +13,13 @@ fn main() -> AppExit {
             ..default()
         })
         .init_state::<Screen>()
+        /*
         .add_plugins(
             ProgressPlugin::new()
                 // Changes to `BevyState<Screen>` will propagate to `Screen`.
                 .with_state_transition(Screen::Loading.bevy(), Screen::Gameplay.bevy()),
         )
+        */
         .add_systems(
             Update,
             Screen::Title.on_update(